Try to reuse system id in post_gen_cleanup
