+++ scripts/__init__.py
@@ -1,0 +1,9 @@
+# ***************************************************************** #
+# (C) Copyright IBM Corporation 2024.                               #
+#                                                                   #
+# The source code for this program is not published or otherwise    #
+# divested of its trade secrets, irrespective of what has been      #
+# deposited with the U.S. Copyright Office.                         #
+# ***************************************************************** #
+# SPDX-License-Identifier: Apache-2.0
+"""This module is for common bash/python scripts"""
